+++ CHANGELOG.rst
@@ -61,4 +61,5 @@ Changelog
 * Add command line options to control testvars gathering
 * Include extra vars (again) as TESTVARS_EXTRA_VARS
 * Provide environment variables as fixtures
-* Start unit tests
+* Add unit tests
+* Refactor TestVars class